+++ src/config-loader.js
@@ -1,5 +1,3 @@
-import { BASE_PATH } from "./variables";
-
 let configData = null;
 let activePromise = null;
 
@@ -13,7 +11,7 @@ const ConfigLoader = {
       if (configData !== null) {
         resolve(configData);
       } else {
-        fetch(`${BASE_PATH}config.json`)
+        fetch("/admin/config.json")
           .then((response) => response.json())
           .then((data) => {
             configData = data;