+++ src/index.jsx
@@ -4,14 +4,13 @@ import { createRoot } from "react-dom/client";
 import { Provider } from "react-redux";
 import { store } from "./redux/store";
 import App from "./app";
-import { BASE_PATH } from "./variables";
 
 const container = document.getElementById("root");
 const root = createRoot(container);
 
 root.render(
   <Provider store={store}>
-    <BrowserRouter basename={BASE_PATH}>
+    <BrowserRouter basename="/admin">
       <App />
     </BrowserRouter>
   </Provider>